move external sources
